+++ lib/streamlit/temporary_directory.py
@@ -15,48 +15,9 @@
 
 import tempfile
 import shutil
-import os
 
-# We provide our own context managers for temporary file and directory
-# that works in Windows as well.
-# As documented in https://github.com/streamlit/streamlit/issues/301,
-# Python `tempfile.NamedTemporaryFile` context manager does not behave
-# as expected in Windows.
-
-
-class TemporaryFile(object):
-    """Temporary file context mananger.
-
-    Create a temporary file that exists whithin the context manager scope.
-    It returns the path to the file. The created file
-    is closed and the user needs to open it to write/read.
-    Wrapper on top of tempfile.mkstemp.
-
-    Parameters
-    ----------
-    suffix : str or None
-        Suffix to the filename.
-    prefix : str or None
-        Prefix to the filename.
-    dir : str or None
-        Enclosing directory.
-
-    """
-
-    def __init__(self, *args, **kwargs):
-        self._args = args
-        self._kwargs = kwargs
-
-    """Context Manager """
-
-    def __enter__(self):
-        fd, self._path = tempfile.mkstemp(*self._args, **self._kwargs)
-        # We close the file descriptor since mkstemp opens it by default.
-        os.close(fd)
-        return self._path
-
-    def __exit__(self, *exec):
-        os.remove(self._path)
+# We provide our own context managers for temporary directory that wraps
+# tempfile.mkdtemp
 
 
 class TemporaryDirectory(object):
