+++ lib/tests/streamlit/temporary_directory_test.py
@@ -18,18 +18,11 @@ import os
 import unittest
 from testfixtures import tempdir
 
-from streamlit.temporary_file import TemporaryFile, TemporaryDirectory
+from streamlit.temporary_directory import TemporaryDirectory
 
 
 class TemporaryFileTest(unittest.TestCase):
-    """Test temp file and directory context manager."""
-
-    @tempdir()
-    def test_temp_file(self, dir):
-        """Test that the file only exists inside the context."""
-        with TemporaryFile(dir=dir.path) as temp_fname:
-            self.assertTrue(os.path.exists(temp_fname))
-        self.assertFalse(os.path.exists(temp_fname))
+    """Test temp directory context manager."""
 
     @tempdir()
     def test_temp_directory(self, dir):
